Move execCmd to `setup_linux.go`
